消費税を小数点以下切り捨てに修正

+++ src/php/app/Http/Controllers/BuyController.php
@@ -42,7 +42,7 @@ class BuyController extends Controller
         }
 
         if ($priceIncludeTax != 0) {
-            $tax =  $priceIncludeTax * 0.08;
+            $tax =  (int)($priceIncludeTax * 0.08);
             $priceIncludeTax += $tax;
         } else {
             $tax = 0;
